src/tests/test_tweet_controller.py

import unittest
from twitter_api import TwitterAPI
from unittest.mock import patch

class TestTweetController(unittest.TestCase):

    @patch('twitter_api.TwitterAPI.post_tweet')
    def test_post_tweet_success(self, mock_post_tweet):
        # Setup mock response
        mock_post_tweet.return_value = {'status': 'success'}

        # Test posting a tweet
        response = TwitterAPI().post_tweet("Hello, world!")
        self.assertIsNotNone(response)
        self.assertEqual(response['status'], 'success')

    @patch('twitter_api.TwitterAPI.post_tweet')
    def test_post_tweet_failure(self, mock_post_tweet):
        # Setup mock response for failure
        mock_post_tweet.return_value = None

        # Test failure in posting a tweet
        response = TwitterAPI().post_tweet("Hello, world!")
        self.assertIsNone(response)

    # Additional test cases...

if __name__ == '__main__':
    unittest.main()
